Initial import of the script to customise Pis from bare Photonic installs
This should take a bare photonic3d install (or raspbian Jessie lite in the case of the 4ktouch) and customise it to spec. It assumes the latest Photonic has been installed to /opt/cwh already. If it hasn't, this script will fail [there's very little error checking in here as it assumes a fairly clean working install.]

It is currently UNTESTED [thus the name]. It has worked prior to being committed, however a lot more functionality has been built in and the amount the script does has increased. It should be a good starting point to develop a better, more resilient script.