Added holding page for touch screens waiting for photonic server to be up and running.
On the kweb browser, it needs to access holdingpage.html locally (i.e. not from the Photonic server/ports. It needs to read it as a local file)

This needs to be supplied with a URL parameter, target set to the appropriate URL for the photonic server.

E.g.:
holdingpage.html?target=http://4kscreen.local:9091/printflow
holdingpage.html?target=http://LCHR.local:9091/printflow

It will constantly try to load a hardcoded image from the printflow images folder on the specified target server. If it fails, it will sleep for 1 second and try again. If it succeeds, it will set the web browser url to be the target location.

Build deployment will need to be updated on LCHR and 4ktouch (and the new standalone box) to support this so that the file (and resources) are copied to the pi's home directory and the kweb script now points to a local copy of holdingpage first and then redirects to the designated URL parameter defined photonic server.