Updating all the 'simple' html files to use js/printflow.js and js/printerconfig.js
Took all the clock stuff and moved it to printflow.js.
As a bonus, all these pages will now warp to printdialogue.html if a
print is started externally to the printer.
They also now use printerconfig.js - this is an auto-generated file by
raspi-setup.sh that tells printflow what device it's installed on.